refactor :: 파일 재구조화 및 코드 리팩토링
