+++ iOS/Sources/Scene/Post/PostListVC/PostDummy.swift
@@ -1,3 +1,22 @@
+import Foundation
+
+struct BirthDay {
+    var username: String = .init()
+    var birthDate: String = .init()
+    let cellType: String = "birthType"
+}
+
+struct Schedule {
+    var color: String = .init()
+    var title: String = .init()
+    var content: String = .init()
+    var owner: String = .init()
+    var date: String = .init()
+    var location: String = .init()
+    var text: String = .init()
+    let cellType: String = "scheduleType"
+}
+
 class Dummies {
     // BirthDaySection
     let birthItem1 = BirthDay.self(username: "Daehee", birthDate: "2022-08-22")
