--- iOS/Sources/Scene/Post/Reactor/PostListVCReactor.swift
@@ -1,16 +1,0 @@
-import Foundation
-import ReactorKit
-
-class PostListVCReactor: Reactor {
-    private let disposeBag: DisposeBag = .init()
-    let initialState: State
-    enum Action {
-    }
-    enum Mutation {
-    }
-    struct State {
-    }
-    init() {
-        initialState = State()
-    }
-}
